% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcr.R
\docType{data}
\name{archery1}
\alias{archery1}
\title{Target archery dataset in the ranking round (used as Phase I)}
\format{
An array of (24 x 2 x 3).
\describe{ 
\item{x-coordinate}{x-coordinate}
\item{y-coordinate}{y-coordinate}
}
}
\description{
It consists of a stage in which the archer shoots 72 arrows in 12 ends of six arrows. 
The information is given in x and y coordinates.
}
\examples{

data(archery1)
str(archery1) ; plot(archery1)
}
\keyword{datasets}
