% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcr.R
\docType{data}
\name{dowel1}
\alias{dowel1}
\title{Dowel pin dataset}
\format{
A data frame with 40 observations on the following 2 variables.
\describe{ 
\item{diameter}{A numeric vector}
\item{length}{A numeric vector}
}
}
\description{
Diameter and length of a manufacturing process of a dowel pin.
}
\examples{

data(dowel1)
str(dowel1) ; plot(dowel1)
}
\keyword{datasets}
