% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Mode}
\alias{Mode}
\title{Compute mode for vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{vector of characters}
}
\value{
Most frequencly occurance unique element
}
\description{
Compute mode for vector
}
\keyword{internal}
