% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mqcd}
\alias{mqcd}
\title{Multivariante Quality Control Data}
\usage{
mqcd(data, data.name = NULL)
}
\arguments{
\item{data}{a matrix or data-frame or array where it should contain data.}

\item{data.name}{a string that specifies the title displayed on the plots.
If not provided is taken from the name of the object data.}
}
\description{
Create an object of class 'mqcd' to perform statistical quality control.
This object be used to plot Multivariate Control Charts.
}
\examples{
library(qcr)
data(dowel1)
str(dowel1)
data.mqcd <- mqcd(dowel1)
str(data.mqcd)
}

