% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqcs.R
\name{mqcs.add}
\alias{mqcs.add}
\alias{mqcs.add.default}
\title{mqcs.add Add a matrix, data.frame or array object with a mqcs object}
\usage{
mqcs.add(x, ...)

\method{mqcs.add}{default}(x, value, ...)
}
\arguments{
\item{x}{Object type mqcs}

\item{...}{arguments to be passed to or from methods.}

\item{value}{Object type data.frame, matrix or array}
}
\description{
This function is used to join two objects of type matrix, data.frame or array and mqcs.
}
