% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mqcs.mcusum}
\alias{mqcs.mcusum}
\alias{mqcs.mcusum.default}
\alias{mqcs.mcusum.mqcd}
\title{Function to plot mcusum chart}
\usage{
mqcs.mcusum(x, ...)

\method{mqcs.mcusum}{default}(x, data.name = NULL, Xmv = NULL, S = NULL,
  k = 0.5, h = 5.5, method = "sw", plot = FALSE, ...)

\method{mqcs.mcusum}{mqcd}(x, Xmv = NULL, S = NULL, k = 0.5, h = 5.5,
  method = "sw", plot = FALSE, ...)
}
\arguments{
\item{x}{an R object (used to select the method). See details.}

\item{...}{arguments passed to or from methods.}

\item{Xmv}{is the mean vector. It is only specified for Phase II or when the parameters of the distribution are known.}

\item{S}{is the sample covariance matrix. It is only used for Phase II or when the parameters of the distribution are known.}

\item{k}{is a constant used in MCUSUM chart. Frequently k = 0.5}

\item{h}{is a constant used in MCUSUM chart. Usually h = 5.5}

\item{method}{Is the method employed to compute the covatiance matrix
in individual observation case. Two methods are used "sw"
for compute according to (Sullivan,Woodall 1996a) and "hm"
by (Holmes,Mergen 1993)}

\item{plot}{a logical value indicating should be plotted.}

\item{data.name}{a string that specifies the title displayed on the plots.
If not provided is taken from the name of the object data.}
}
\description{
This function is used to compute statistics required by the mcusum chart.
}
\examples{
##
##  Continuous data
##
library(qcr)
data(dowel1)
str(dowel1)
data.mqcd <- mqcd(dowel1)
res.mqcs <- mqcs.mcusum(data.mqcd)
summary(res.mqcs)
plot(res.mqcs, title =" MCUSUM Control Chart for dowel1")
}
\author{
Edgar Santos-Fernandez
}

