% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mqcs}
\alias{mqcs}
\alias{print.mqcs}
\alias{summary.mqcs}
\title{Create an object of class 'mqcs' to perform statistical quality control.
This function is used to compute statistics required for to plot Multivariate Control Charts}
\usage{
mqcs(x, method = "sw", ...)
}
\arguments{
\item{x}{Object mqcd (Multivariante Quality Control Data)}

\item{method}{Is the method employed to compute the covatiance matrix
in individual observation case. Two methods are used "sw"
for compute according to (Sullivan,Woodall 1996a) and "hm"
by (Holmes,Mergen 1993)}

\item{...}{arguments passed to or from methods.}
}
\description{
Create an object of class 'mqcs' to perform statistical quality control.
This function is used to compute statistics required for to plot Multivariate Control Charts
}

