% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npqcs.R
\name{npqcs.add}
\alias{npqcs.add}
\alias{npqcs.add.default}
\title{npqcs.add Add a matrix, data.frame or array object with a npqcs object}
\usage{
npqcs.add(x, ...)

\method{npqcs.add}{default}(x, value, ...)
}
\arguments{
\item{x}{Object type npqcs}

\item{...}{Arguments to be passed to or from methods.}

\item{value}{Object type data.frame, matrix or array.}
}
\description{
This function is used to join two objects of type matrix, data.frame or array and npqcs.
}
