% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npqcs.R
\name{npqcs}
\alias{npqcs}
\alias{summary.npqcs}
\alias{print.npqcs}
\title{Statistical Quality Control Object}
\usage{
npqcs(x, method = c("Tukey", "Liu", "Mahalanobis", "RP", "LD"), ...)
}
\arguments{
\item{x}{Object npqcd (Non Parametric Multivariante Quality Control Data)}

\item{method}{Character string which determines the depth function used. 
\code{method} can be "Tukey" (the default), "Liu", "Mahalanobis", "RP" Random Project or "LD" Likelihood depth.}

\item{...}{Arguments passed to or from methods.}
}
\description{
Create an object of class 'npqcs' to perform statistical quality control.
This function is used to compute statistics required to plot 
Non Parametric Multivariate Control Charts.
}
