% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npqcs.S.r
\name{npqcs.S}
\alias{npqcs.S}
\alias{npqcs.S.default}
\alias{npqcs.S.npqcd}
\title{Function to plot the S chart}
\usage{
npqcs.S(x, ...)

\method{npqcs.S}{default}(
  x,
  G,
  data.name = NULL,
  limits = NULL,
  method = c("Tukey", "Liu", "Mahalanobis", "RP", "LD"),
  alpha = 0.01,
  plot = FALSE,
  standardize = FALSE,
  ...
)

\method{npqcs.S}{npqcd}(
  x,
  data.name,
  limits = NULL,
  method = c("Tukey", "Liu", "Mahalanobis", "RP", "LD"),
  alpha = 0.01,
  plot = FALSE,
  standardize = F,
  ...
)
}
\arguments{
\item{x}{An object of class "npqcd".}

\item{...}{Arguments passed to or from methods.}

\item{G}{The x as a matrix, data frame or list. If it is a matrix or data frame,
then each row is viewed as one multivariate observation.}

\item{data.name}{A string that specifies the title displayed on the plots. 
If not provided it is taken from the name of the object \code{x}.}

\item{limits}{A two-value vector specifying the control limits lower and central.}

\item{method}{Character string which determines the depth function used. 
\code{method} can be "Tukey" (the default), "Liu", "Mahalanobis", "RP" Random Project or "LD" Likelihood depth.}

\item{alpha}{It is the significance level (by default \code{alpha} = 0.01)}

\item{plot}{Logical value. If \code{TRUE} a S chart should be plotted.}

\item{standardize}{A logical value indicating data should be standardized.}
}
\description{
This function is used to compute statistics required by the Non Parametric S chart.
}
\examples{
\dontrun{
##
##  Continuous data 
##
set.seed(12345)
mu<-c(0,0)
Sigma<- matrix(c(1,0,0,1),nrow = 2,ncol = 2)
u <- c(2,2)
S <- matrix(c(4,0,0,4),nrow = 2,ncol = 2)
G <- rmvnorm(540, mean = mu, sigma = Sigma)
x<- rmvnorm(40,mean=u,sigma = S)
x.a <- rbind(G[501:540,],x)
M <- G[1:500,]
data.npqcd <- npqcd(x.a,M)
res.npqcs <- npqcs.S(data.npqcd,method = "Liu", alpha=0.05)
summary(res.npqcs)
plot(res.npqcs,title =" S Control Chart")}
}
\references{
Regina Liu (1995)
}
