% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcr.R
\docType{data}
\name{pcmanufact}
\alias{pcmanufact}
\title{Personal computer manufacturer data}
\format{
A data frame with 10 observations on the following 2 variables.
\describe{
\item{x}{number of nonconformities (inspection units)}
\item{sample}{sample ID}
\item{size}{number of computers inspected}
}
}
\description{
A personal computer manufacturer counts the number of nonconformities per
unit on the final assembly line. He collects data on 20 samples of 5
computers each.
}
\examples{

data(pcmanufact)
summary(pcmanufact)
plot(pcmanufact$x/pcmanufact$size, type="b")
}
\references{
Montgomery, D.C. (1991) \emph{Introduction to Statistical
Quality Control}, 2nd ed, New York, John Wiley & Sons, pp. 181--182
}
\keyword{datasets}
