% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcr.R
\docType{data}
\name{plates}
\alias{plates}
\title{Vickers hardness data}
\format{
A data frame with 250 observations on the following 2 variables:
\describe{
  \item{hardness}{A quantitative variable that measure Vickers hardness corresponding to different
   areas of artificial stone.}
  \item{sample}{Sample ID}
}
}
\description{
A known chemical company is developing a patent for 
a new variant of artificial stone composed mostly of quartz ( 93wt %) 
and polyester resin . This company is launching a pilot plant where 
it begins to produce plates of this material to industry scale. In order 
to measure the degree of product homogeneity, 50 samples were taken, 
performed 5 measurements per plate corresponding to different areas 
of artificial stone Vickers hardness
}
\examples{

data(plates)
attach(plates)
summary(plates)
plot(hardness, type="b")
detach(plates)
}
\keyword{datasets}
