% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fdqcs.R
\name{plot.fdqcs.rank}
\alias{plot.fdqcs.rank}
\title{Plot method for 'fdqcs.depth' objects
Generic function for plotting charts of object of class 
'fdqcs.rank' to perform statistical quality control.}
\usage{
\method{plot}{fdqcs.rank}(
  x,
  title.fdata = NULL,
  title.rank = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  draw.control = NULL,
  ...
)
}
\arguments{
\item{x}{Object fdqcs.depth}

\item{title.fdata}{an overall title for the fdata plot}

\item{title.rank}{an overall title for the depth plot}

\item{xlab}{a title for the x axis}

\item{ylab}{a title for the y axis}

\item{draw.control}{ist that it specifies the col, lty and lwd for objects: fdataobj, statistic, IN and OUT.}

\item{...}{arguments to be passed to or from methods.}
}
\description{
Plot method for 'fdqcs.depth' objects
Generic function for plotting charts of object of class 
'fdqcs.rank' to perform statistical quality control.
}
