% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fdqcs.depth.R
\name{plot.fdqcs.depth}
\alias{plot.fdqcs.depth}
\title{Plot method for 'fdqcs.depth' objects}
\usage{
\method{plot}{fdqcs.depth}(
  x,
  title.fdata = NULL,
  title.depth = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  draw.control = NULL,
  ...
)
}
\arguments{
\item{x}{Object fdqcs.depth}

\item{title.fdata}{An overall title for the fdata plot.}

\item{title.depth}{An overall title for the depth plot.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{col}{The color for curves.}

\item{draw.control}{It specifies the col, lty and lwd for objects: 
fdataobj, statistic, IN and OUT.}

\item{...}{Arguments to be passed to or from methods.}
}
\description{
Generic function for plotting charts of object of class 'fdqcs.depth' to perform statistical 
quality control.
}
