% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{presion}
\alias{presion}
\title{Level of presion data}
\format{A data frame with 180 observations on the following 3 variables:
\describe{
  \item{presion}{presion level}
  \item{sample}{sample id}
  \item{measur}{pressures measured: 0.1 y 10 MPa}
}}
\description{
A shipyard of recreational boats manufacturing,
intended to optimize and control the mechanical properties hull yacht models.
This has made a study in which the modulus of elasticity tensile strength of the epoxy resin
(polymer) used, after applying different curing pressures measured: 0.1 y 10 MPa.
60 subsamples composed of three measurements taken on the same day are taken.
}
\examples{
data(presion)
attach(presion)
summary(presion)
plot(presion$presion, type="b")
detach(presion)
}
\keyword{datasets}

