% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{QC_cluster_class}
\alias{QC_cluster_class}
\title{Checks each cluster for homogeneity of class level annotations. (Prev: QC2)}
\usage{
QC_cluster_class(rnaseq.data, class.col, class.cutoff = 0.75)
}
\arguments{
\item{rnaseq.data}{Seurat object}

\item{class.col}{Column of suerat object that contains \code{class} annotations}

\item{class.cutoff}{Clusters with less than class.cutoff homogeneity are flagged}
}
\value{
cluster.class.flag
}
\description{
Checks each cluster for homogeneity of class level annotations. (Prev: QC2)
}
