% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{QC_label_conf}
\alias{QC_label_conf}
\title{Check for confidence in cell annotation}
\usage{
QC_label_conf(rnaseq.data, label.confidence, label.confidence.threshold)
}
\arguments{
\item{rnaseq.data}{Seurat object}

\item{label.confidence}{Confidence value per cell}

\item{label.confidence.threshold}{Threshold to determine low confidence and acceptable annotations}
}
\value{
label.conf.flag
}
\description{
Check for confidence in cell annotation
}
