% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{qcr}
\alias{qcr}
\alias{qcr-package}
\title{Quality Control and Reability}
\description{
Quality control and reliability
}
\details{
This package allows to generate Shewhart-type charts and to obtain
numerical results of interest to the quality control of a process
(involving  continuous, attribute or count data).
This package provides basic functionality for univariable and multivariable
quality control analysis, including: xbar, xbar-one, S, R, n, np, c, g, ewna, cusum,
mewna, mcusum  and T2 charts.
}

