% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qcs.add}
\alias{qcs.add}
\alias{qcs.add.default}
\title{qcs.add Add a data.frame object with a qcs object}
\usage{
qcs.add(x, ...)

\method{qcs.add}{default}(x, value, var.index = NULL, sample.index = NULL,
  covar.index = NULL, ...)
}
\arguments{
\item{x}{Object type qcs}

\item{value}{Object type data.frame}

\item{var.index}{a scalar with the column number corresponding the observed data for
the variable (the variable quality).  Alternativelly can be a string with the
name of the quality variable.}

\item{sample.index}{a scalar with the column number corresponding the index each
group (sample).}

\item{covar.index}{optional. A scalar or numeric vector with the column number(s)
corresponding to the covariate(s). Alternativelly can be a character vector with
the names of the covariates.}

\item{...}{arguments to be passed to or from methods.}
}
\description{
This function is used to join two objects of type data.frame and qcs.
}

