% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcs.c.r
\name{qcs.c}
\alias{qcs.c}
\alias{qcs.c.default}
\alias{qcs.c.qcd}
\title{Function to plot Shewhart c chart}
\usage{
qcs.c(x, ...)

\method{qcs.c}{default}(
  x,
  var.index = 1,
  sample.index = 2,
  covar.index = NULL,
  covar.names = NULL,
  data.name = NULL,
  sizes = NULL,
  center = NULL,
  conf.nsigma = 3,
  limits = NULL,
  plot = FALSE,
  ...
)

\method{qcs.c}{qcd}(x, center = NULL, conf.nsigma = 3, limits = NULL, plot = FALSE, ...)
}
\arguments{
\item{x}{An object of class "qcd".}

\item{...}{Arguments passed to or from methods.}

\item{var.index}{A scalar with the column number corresponding to the observed data for
the variable (the variable quality).  Alternativelly can be a string with the
name of the quality variable.}

\item{sample.index}{A scalar with the column number corresponding to the index fo each
group (sample).}

\item{covar.index}{Optional. A scalar or numeric vector with the column number(s)
corresponding to the covariate(s). Alternativelly it can be a character vector with
the names of the covariates.}

\item{covar.names}{Optional. A string or vector of strings indicating the names for the
covariate columns.  Only valid if there is more than one column of data. By
default, takes the names from the original object.}

\item{data.name}{A string specifying the name of the variable which appears on the
plots. If not provided, it is taken from the object given as data.}

\item{sizes}{Optional. A value or a vector of values specifying the sample sizes
associated with each group. For continuous data, the sample sizes are obtained 
counting the non-\code{NA} elements of the \code{sample.index} vector. 
For attribute variable the argument sizes is required.}

\item{center}{A value specifying the center of group statistics or the
''target'' value of the process.}

\item{conf.nsigma}{A numeric value used to compute control limits, specifying the
number of standard deviations (if \code{conf.nsigma} > 1) or the confidence level (if 0
< \code{conf.nsigma} < 1).}

\item{limits}{A two-value vector specifying control limits.}

\item{plot}{Logical value. If \code{TRUE} a c chart should be plotted.}
}
\description{
This function is used to compute statistics required by the c chart.
}
\examples{
library(qcr)
data(circuit)
attach(circuit)
str(circuit)
datos <- circuit
datos$sample <- 1:length(datos$x)
str(datos)
sizes <- datos[,2]

datos.qcd <- qcd(data = datos, var.index = 1,sample.index = 2,
                sizes = size, type.data = "atributte")
res.qcs <- qcs.c(datos.qcd)
summary(res.qcs)
plot(res.qcs)

}
