% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seurat_cluster}
\alias{seurat_cluster}
\title{Seurat clustering}
\usage{
seurat_cluster(rnaseq.data, resolution.value)
}
\arguments{
\item{rnaseq.data}{Seurat object}

\item{resolution.value}{Seurat clustering resolution parameter}
}
\value{
Seurat object with PCA, UMAP, clustering
}
\description{
This function takes a seurat object and performs dim. reduction and clustering.
}
\keyword{internal}
