% Generated by roxygen2 (4.0.1): do not edit by hand
\name{state.control}
\alias{state.control}
\title{This function is used for the state of the process is under control.
Were removed from the sample observations that violate the rules of a process under control}
\usage{
state.control(x)
}
\arguments{
\item{x}{Object qcs (Quality Control Statitical)}
}
\description{
This function is used for the state of the process is under control.
Were removed from the sample observations that violate the rules of a process under control
}
\examples{
##
##  Continuous data
##
library(qcr)
data(pistonrings)
str(pistonrings)
pistonrings.qcd<-qcd(pistonrings)

class(pistonrings.qcd)

res.qcs <- qcs.xbar(pistonrings.qcd)
plot(res.qcs,title="Control Chart Xbar for pistonrings I")
summary(res.qcs)

res.qcd <- state.control(res.qcs)
res.qcs <- qcs.xbar(res.qcd)
plot(res.qcs,title="Control Chart Xbar for pistonrings II")
summary(res.qcs)

res.qcd <- state.control(res.qcs)
res.qcs <- qcs.xbar(res.qcd)
plot(res.qcs,title="Control Chart Xbar for pistonrings III")
summary(res.qcs)
}

