% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcr.R
\docType{data}
\name{counters}
\alias{counters}
\title{The performance of the counters data}
\format{
A data frame with 180 observations on the following 3 variables:
\describe{
  \item{error}{the measurement error of the counters (Error:  (Real Volume - Measured Volume)/Real Volume)}
  \item{sample}{sample id}
  \item{brand}{brands of providers of counters}
}
}
\description{
A water company from A Corunia wants to control the performance 
of the counters installed throughout the city.
60 subsamples are taken each one composed by 3 measurements made 
by the counters of the same antiquity (10 years) and caliber, 
in a period of 5 years. Taking into account that there are two brands 
or providers of counters
}
\examples{

data(counters)
attach(counters)
summary(counters)
plot(error, type="b")
detach(counters)
}
\keyword{datasets}
