% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcr.R
\docType{data}
\name{counters}
\alias{counters}
\title{The performance of the counters data}
\format{
A data frame with 180 observations on the following 3 variables:
\describe{
  \item{error}{The measurement error of the counters (Error:  (Real Volume - Measured Volume)/Real Volume)}
  \item{sample}{Sample ID}
  \item{brand}{Brands of providers of counters}
}
}
\description{
A water supply company wants to control the performance of 
the water counters installed throughout a city. For this purpose, 
60 rational samples have been taken, each one composed by 3 measurements, 
from the same age (10 years) and caliber water counters corresponding to 
two different brands, and during a period of 5 years. This dataset is
based on a study case of A Coruña’s water supply company, Empresa 
Municipal de Aguas de La Coruña (Emalcsa).
}
\examples{

data(counters)
attach(counters)
summary(counters)
plot(error, type="b")
detach(counters)
}
\keyword{datasets}
