% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcr.R
\docType{data}
\name{employment}
\alias{employment}
\title{Level of employment data}
\format{
A data frame with 288 observations on the following 3 variables:
\describe{
  \item{occupantion}{the amount of occupants in terms of percentage}
  \item{sample}{sample id}
  \item{hemisphere}{Hemisphere}
}
}
\description{
A Spaniard-Argentinian hotel company wants to control the 
level of occupancy (measured in %) in their establishments through the
application of a continuous control. For this purpose, 48 subsamples have
been taken from six hotels corresponding to two different countries.
}
\examples{

data(employment)
attach(employment)
summary(employment)
boxplot(occupantion ~ hemisphere)
plot(occupantion, type="b")
detach(employment)
}
\keyword{datasets}
