% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqcd.R
\name{mqcd}
\alias{mqcd}
\title{It creates a data object to be used in Multivariante Quality Control}
\usage{
mqcd(data, data.name = NULL)
}
\arguments{
\item{data}{A matrix or data-frame or array where it should contain data.}

\item{data.name}{A string that specifies the title displayed on the plots. 
If not provided it is taken from the name of the object's data.}
}
\description{
Create an object of class 'mqcd' to perform statistical quality control.
This object is used to plot Multivariate Control Charts.
}
\examples{
library(qcr)
data(dowel1)
str(dowel1)
data.mqcd <- mqcd(dowel1)
str(data.mqcd)
}
