% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqcs.R
\name{mqcs}
\alias{mqcs}
\alias{summary.mqcs}
\alias{print.mqcs}
\title{It computes statistics to be used in Multivariante Quality Control}
\usage{
mqcs(x, method = "sw", ...)
}
\arguments{
\item{x}{Object mqcd (Multivariante Quality Control Data)}

\item{method}{Is the method employed to compute the covariance matrix
in individual observation case. Two methods are used "sw" 
for compute according to (Sullivan,Woodall 1996a) and "hm" 
by (Holmes,Mergen 1993)}

\item{...}{Arguments passed to or from methods.}
}
\description{
Create an object of class 'mqcs' to perform statistical quality control.
This function is used to compute statistics required to plot Multivariate Control Charts
}
