% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcr.R
\docType{data}
\name{oxidation}
\alias{oxidation}
\title{Oxidation Onset Temperature}
\format{
A data frame with 250 observations on the following 2 variables:
\describe{
  \item{OOT}{That is a quantitative variable that controls the quality of oil.}
  \item{sample}{sample id}
}
}
\description{
This database contains information about the level of purity 
of each batch of Picual varities. Then we have the type of oil 
by measuring the Oxidation Onset Temperature. 
We have 50 subsamples of oil with their temperature to oxide.
}
\examples{

data(oxidation)
attach(oxidation)
summary(oxidation)
plot(OOT, type="b")
detach(oxidation)
}
\keyword{datasets}
