% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcr.R
\docType{data}
\name{pcmanufact}
\alias{pcmanufact}
\title{Personal computer manufacturer data}
\format{
A data frame with 20 observations on the following 3 variables.
\describe{
\item{x}{A quantitative variable that measures the number of nonconformities in each sample.}
\item{sample}{Sample ID}
\item{size}{Number of computers inspected.}
}
}
\description{
A personal computer manufacturer counts the number of nonconformities per
unit on the final assembly line. He collects data on 20 samples of 5
computers each.
}
\examples{

data(pcmanufact)
summary(pcmanufact)
plot(pcmanufact$x/pcmanufact$size, type="b")
}
\references{
Montgomery, D.C. (1991) \emph{Introduction to Statistical
Quality Control}, 2nd ed, New York, John Wiley & Sons, pp. 181--182
}
\keyword{datasets}
