% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fdqcs.rank.R
\name{plot.fdqcs.rank}
\alias{plot.fdqcs.rank}
\title{Plot method for 'fdqcs.rank' objects}
\usage{
\method{plot}{fdqcs.rank}(
  x,
  title.fdata = NULL,
  title.rank = NULL,
  xlab = NULL,
  ylab = NULL,
  col = NULL,
  draw.control = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class fdqcs.rank}

\item{title.fdata}{An overall title for the fdata plot.}

\item{title.rank}{An overall title for the depth plot.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{draw.control}{It specifies the col, lty and lwd for objects: 
fdataobj, statistic, IN and OUT.}

\item{...}{Arguments to be passed to or from methods.}
}
\description{
Generic function for plotting charts of object of class 'fdqcs.rank' to 
perform statistical quality control.
}
