% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mqcs.R
\name{plot.mqcs}
\alias{plot.mqcs}
\alias{plot.mqcs.t2}
\alias{plot.mqcs.mcusum}
\alias{plot.mqcs.mewma}
\title{Plot method for 'mqcs' objects}
\usage{
\method{plot}{mqcs}(x, title, subtitle, xlab, ylab, ylim, ...)

\method{plot}{mqcs.t2}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ...
)

\method{plot}{mqcs.mcusum}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ...
)

\method{plot}{mqcs.mewma}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{Object mqcs (Multivarite Quality Control Statical)}

\item{title}{An overall title for the plot.}

\item{subtitle}{A sub title for the plot.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{ylim}{The y limits of the plot.}

\item{...}{Arguments to be passed to or from methods.}
}
\description{
Generic function for plotting Multivarite charts of object of class 'mqcs' 
to perform statistical quality control.
}
