% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.qcs.R
\name{plot.qcs}
\alias{plot.qcs}
\alias{plot.qcs.xbar}
\alias{plot.qcs.S}
\alias{plot.qcs.R}
\alias{plot.qcs.one}
\alias{plot.qcs.p}
\alias{plot.qcs.np}
\alias{plot.qcs.c}
\alias{plot.qcs.u}
\alias{plot.qcs.ewma}
\alias{plot.qcs.cusum}
\title{Function to create a plotting 'qcs' object}
\usage{
\method{plot}{qcs}(
  x,
  title,
  subtitle,
  xlab,
  ylab,
  ylim,
  center.nominal = NULL,
  limits.specification = NULL,
  limits.alert = NULL,
  type.data = c("continuous", "atributte", "dependence"),
  ...
)

\method{plot}{qcs.xbar}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

\method{plot}{qcs.S}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

\method{plot}{qcs.R}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

\method{plot}{qcs.one}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

\method{plot}{qcs.p}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

\method{plot}{qcs.np}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

\method{plot}{qcs.c}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

\method{plot}{qcs.u}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  conf.nsigma.alert = NULL,
  center.nominal = NULL,
  limits.specification = NULL,
  ...
)

\method{plot}{qcs.ewma}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ...
)

\method{plot}{qcs.cusum}(
  x,
  title = NULL,
  subtitle = NULL,
  xlab = NULL,
  ylab = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{Object qcs (Quality Control Statical).}

\item{title}{An overall title for the plot.}

\item{subtitle}{A sub title for the plot.}

\item{xlab}{A title for the x axis.}

\item{ylab}{A title for the y axis.}

\item{ylim}{The y limits of the plot.}

\item{center.nominal}{A value specifying the 
center of group statistics or the "target" value 
of the process.}

\item{limits.specification}{A two-value vector 
specifying control limits.}

\item{limits.alert}{A two-value vector 
specifying control alert limits.}

\item{type.data}{A string specifying the type of data.}

\item{...}{Arguments to be passed to or from methods.}

\item{conf.nsigma.alert}{A numeric value used to compute control limits, specifying the
number of standard deviations (if \code{conf.nsigma} > 1) or the confidence level (if 0
< \code{conf.nsigma} < 1).}

\item{conf.nsigma}{A numeric value used to compute control limits, specifying the
number of standard deviations (if \code{conf.nsigma} > 1) or the confidence level (if 0
< \code{conf.nsigma} < 1).}
}
\description{
Generic function for plotting Shewhart charts of object of class 'qcs' to perform statistical 
quality control.
}
