% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcr.R
\name{QCR}
\alias{QCR}
\title{QCR}
\usage{
QCR(
  rnaseq.data,
  class.col,
  class.cutoff,
  neuron.class.id,
  neuron.cutoff,
  non.neuron.class.id,
  non.neuron.cutoff,
  class.col.score,
  subclass.col.score,
  label.confidence.threshold = 0.7,
  resolution.value = 10
)
}
\arguments{
\item{rnaseq.data}{Seurat object}

\item{class.col}{Column of suerat object that contains \code{class} annotations}

\item{class.cutoff}{Clusters with less than class.cutoff homogeneity are flagged}

\item{neuron.class.id}{Name of neuronal cells within the \code{class} annotations}

\item{neuron.cutoff}{Proportion of top/bottom clusters to flag for neuron clusters}

\item{non.neuron.class.id}{Name of non-neuronal cells within the \code{class} annotations}

\item{non.neuron.cutoff}{Proportion of top/bottom clusters to flag for non-neuron clusters}

\item{class.col.score}{Confidence value per cell for class annotation}

\item{subclass.col.score}{Confidence value per cell for subclass annotation}

\item{label.confidence.threshold}{Threshold to determine low confidence and acceptable annotations}

\item{resolution.value}{Seurat clustering resolution parameter if \code{seurat_clusters} is not defined in rnaseq.data}
}
\value{
data.frame with QCR flags
}
\description{
This function computes various QC flags for a seurat object
}
