% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcr.R
\docType{package}
\name{qcr}
\alias{qcr}
\alias{qcr-package}
\title{Quality Control Review}
\description{
Quality Control Review
}
\details{
Allows to generate Shewhart-type charts and to obtain 
numerical results of interest for a process quality control
(involving  continuous, attribute or count data).
This package provides basic functionality for univariable and multivariable 
quality control analysis, including: xbar, xbar-one, S, R, n, np, c, g, ewna, cusum, 
mewna, mcusum  and T2 charts. Additionally have non parametric control  charts multivariate. 
Parametric and nonparametric Process Capability Indices.
}

