% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcs.r
\name{qcs.add}
\alias{qcs.add}
\alias{qcs.add.default}
\title{qcs.add Add a data.frame object with a qcs object}
\usage{
qcs.add(x, ...)

\method{qcs.add}{default}(
  x,
  value,
  var.index = NULL,
  sample.index = NULL,
  covar.index = NULL,
  ...
)
}
\arguments{
\item{x}{Object type qcs}

\item{...}{Arguments to be passed to or from methods.}

\item{value}{Object type data.frame}

\item{var.index}{A scalar with the column number corresponding to the observed data for
the variable (the variable quality).  Alternatively it can be a string with the
name of the quality variable.}

\item{sample.index}{A scalar with the column number corresponding the index each
group (sample).}

\item{covar.index}{Optional. A scalar or numeric vector with the column number(s)
corresponding to the covariate(s). Alternatively can be a character vector with
the names of the covariates.}
}
\description{
This function is used to join two objects of type data.frame and qcs.
}
